gak eroh

+++ database/migrations/2024_09_05_035319_create_enrollments_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('course_id')->constrained()->onDelete('cascade');
             $table->timestamp('enrollment_date')->useCurrent(); // Automatically set enrollment time
             $table->string('payment_method')->nullable(); // Payment method (optional)
-            $table->decimal('payable_amount', 10, 2)->nullable(); // Amount to be paid (optional)
+            $table->double('payable_amount')->nullable(); // Amount to be paid (optional)
             $table->enum('status', ['pending', 'active', 'canceled'])->default('pending'); // Enrollment status
             $table->dateTime('exp_time')->nullable(); // Cancellation time (optional)
             $table->boolean('is_lifetime')->default(false); // Lifetime access
